Display auto-generated fundamental domains!
The most significant addition is the code that builds the triangle
tree. I tested the auto-generation routine on every dessin we have a
handmade fundamental domain for. In every case, the auto-generated
domain is similar or identical to the handmade one. The domains I
tested are listed in the `tested` file.